+++ packages/editor/src/lib/editor/shapes/VideoShape/videoShapeTypes.ts
@@ -1,0 +1,14 @@
+import { TLAssetId, TLBaseShape } from '@tldraw/tlschema'
+
+/** @public */
+export type TLVideoShapeProps = {
+	w: number
+	h: number
+	time: number
+	playing: boolean
+	url: string
+	assetId: TLAssetId | null
+}
+
+/** @public */
+export type TLVideoShape = TLBaseShape<'video', TLVideoShapeProps>
